fix: Refactor and structure infisical client
