+++ src/core/configs.py
@@ -12,7 +12,7 @@ from pydantic_settings import (
     SettingsConfigDict,
 )
 
-from src.core.constants import CommonPaths
+from src.common.constants import CommonPaths
 from src.core.utils import FileUtil
